refactor: reorganize beatree update modules
